# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FTP VPN Service API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginRequest {
  /** User access key (format: XXXX-XXXX-XXXX-XXXX) */
  accessKey: string;
}

export interface User {
  id: number;
  accessKey: string;
  createdAt: string;
  lastSeenAt?: string;
}

export interface LoginResponse {
  success: boolean;
  user: User;
}

export interface SubscriptionPlan {
  id: number;
  name: string;
  durationDays: number;
  priceRub: number;
  priceCrypto: number;
  cryptoCurrency: string;
  maxDevices: number;
  features: string[];
}

export type SubscriptionStatus =
  (typeof SubscriptionStatus)[keyof typeof SubscriptionStatus];

export const SubscriptionStatus = {
  active: "active",
  expired: "expired",
  pending: "pending",
} as const;

export interface Subscription {
  id: number;
  planId: number;
  planName: string;
  status: SubscriptionStatus;
  startedAt: string;
  expiresAt: string;
  daysLeft: number;
}

export interface VpnConfig {
  id: number;
  name: string;
  serverLocation: string;
  serverFlag: string;
  protocol: string;
  createdAt: string;
  isActive: boolean;
}

export interface GenerateConfigRequest {
  name: string;
  serverLocation: string;
}

export interface ConfigDownload {
  id: number;
  filename: string;
  /** WireGuard config file content */
  content: string;
}

export interface CreatePaymentRequest {
  planId: number;
}

export interface CryptoPaymentResponse {
  paymentId: string;
  invoiceUrl: string;
  amount: number;
  currency: string;
  expiresAt: string;
}

export interface SbpPaymentResponse {
  paymentId: string;
  qrUrl: string;
  amount: number;
  bankName: string;
  expiresAt: string;
}

export type PaymentStatusStatus =
  (typeof PaymentStatusStatus)[keyof typeof PaymentStatusStatus];

export const PaymentStatusStatus = {
  pending: "pending",
  paid: "paid",
  expired: "expired",
  failed: "failed",
} as const;

export interface PaymentStatus {
  paymentId: string;
  status: PaymentStatusStatus;
  method: string;
}

export interface UserStats {
  totalConfigs: number;
  activeConfigs: number;
  subscriptionStatus: string;
  daysUntilExpiry: number;
  accountAgedays?: number;
}