# lib/api-zod/src/generated/types/cryptoPaymentResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FTP VPN Service API
 * OpenAPI spec version: 0.1.0
 */

export interface CryptoPaymentResponse {
  paymentId: string;
  invoiceUrl: string;
  amount: number;
  currency: string;
  expiresAt: Date;
}