# lib/api-zod/src/generated/types/generateConfigRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FTP VPN Service API
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateConfigRequest {
  name: string;
  serverLocation: string;
}