# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FTP VPN Service API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  accessKey: string;
  createdAt: Date;
  lastSeenAt?: Date;
}