# lib/api-zod/src/generated/types/vpnConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FTP VPN Service API
 * OpenAPI spec version: 0.1.0
 */

export interface VpnConfig {
  id: number;
  name: string;
  serverLocation: string;
  serverFlag: string;
  protocol: string;
  createdAt: Date;
  isActive: boolean;
}